Moves Logger to log_collector.rs
