+++ program-runtime/src/invoke_context.rs
@@ -2,9 +2,8 @@ use crate::{
     ic_logger_msg, ic_msg,
     instruction_processor::{ExecuteDetailsTimings, Executor, Executors, PreAccount},
     instruction_recorder::InstructionRecorder,
-    log_collector::LogCollector,
+    log_collector::{LogCollector, Logger},
 };
-use log::*;
 use solana_sdk::{
     account::{AccountSharedData, ReadableAccount},
     compute_budget::ComputeBudget,
@@ -44,61 +43,6 @@ impl ComputeMeter {
     pub fn new_ref(remaining: u64) -> Rc<RefCell<Self>> {
         Rc::new(RefCell::new(Self { remaining }))
     }
-}
-
-/// Log messages
-pub struct Logger {
-    log_collector: Option<Rc<LogCollector>>,
-}
-impl Logger {
-    /// Is logging enabled
-    pub fn log_enabled(&self) -> bool {
-        log_enabled!(log::Level::Info) || self.log_collector.is_some()
-    }
-    /// Log a message.
-    ///
-    /// Unless explicitly stated, log messages are not considered stable and may change in the
-    /// future as necessary
-    pub fn log(&self, message: &str) {
-        debug!("{}", message);
-        if let Some(log_collector) = &self.log_collector {
-            log_collector.log(message);
-        }
-    }
-    /// Construct a new one
-    pub fn new_ref(log_collector: Option<Rc<LogCollector>>) -> Rc<RefCell<Self>> {
-        Rc::new(RefCell::new(Self { log_collector }))
-    }
-}
-
-/// Convenience macro to log a message with an `Rc<RefCell<Logger>>`
-#[macro_export]
-macro_rules! ic_logger_msg {
-    ($logger:expr, $message:expr) => {
-        if let Ok(logger) = $logger.try_borrow_mut() {
-            if logger.log_enabled() {
-                logger.log($message);
-            }
-        }
-    };
-    ($logger:expr, $fmt:expr, $($arg:tt)*) => {
-        if let Ok(logger) = $logger.try_borrow_mut() {
-            if logger.log_enabled() {
-                logger.log(&format!($fmt, $($arg)*));
-            }
-        }
-    };
-}
-
-/// Convenience macro to log a message with an `InvokeContext`
-#[macro_export]
-macro_rules! ic_msg {
-    ($invoke_context:expr, $message:expr) => {
-        $crate::ic_logger_msg!($invoke_context.get_logger(), $message)
-    };
-    ($invoke_context:expr, $fmt:expr, $($arg:tt)*) => {
-        $crate::ic_logger_msg!($invoke_context.get_logger(), $fmt, $($arg)*)
-    };
 }
 
 pub struct InvokeContextStackFrame<'a> {
